Updating NetworkTransform with final version
