--- com.unity.multiplayer.mlapi/Prototyping/NetworkTransformV2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eab53a020970f4cbb9ddfe3792b069be
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
